% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMortTable.R
\name{getTable}
\alias{getTable}
\title{Getting Data from Mortality Tables hosted on http://mort.soa.org}
\usage{
getTable(num)
}
\arguments{
\item{num}{The number of the table you want to access in the website}
}
\value{
A list or dataframe contains various information in the table you have asked
}
\description{
mortalitytable access to the website that contains mortality table and grab the information for you.
}
