% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMortTable.R
\name{getTableInfo}
\alias{getTableInfo}
\title{Getting search result from Mortality Tables hosted on http://mort.soa.org}
\usage{
getTableInfo(searchInfo, shortTname = TRUE)
}
\arguments{
\item{searchInfo}{Search information which is the result of function makeSearchInfo}

\item{shortTname}{Set this value as TRUE will make the short table name, otherwise it will give you the table name and the other description too.}
}
\value{
A list or dataframe contains various information in the table you have asked
}
\description{
mortalitytable access to the website that contains mortality table and grab the information for you.
}
