% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMortTable.R
\name{makeSearchInfo}
\alias{makeSearchInfo}
\title{Making the search form for the function "getTableInfo"}
\usage{
makeSearchInfo(
  keywords = "",
  nation = "-1",
  tableUsage = "-1",
  tableType = "-1",
  maxRow = 100,
  sortIndex = "TableIdentity",
  sortDirection = "asc"
)
}
\arguments{
\item{keywords}{Search Key words}

\item{nation}{Choose nation options: United States of America, Argentia, Australia etc.}

\item{tableUsage}{Table usage options: ADB, AD&D, Annuitant Mortality, Claim Cost etc.}

\item{tableType}{Table type options: Aggregate, Attained Age, Continuance etc.}

\item{maxRow}{maxRow controls the maximum number of result you will get.}

\item{sortIndex}{sort the result using the preferred variable}

\item{sortDirection}{Choose "asc" or "desc" for sorting direction of the search result}
}
\value{
A list contains search information you want to attain
}
\description{
https://mort.soa.org/ search page has the various options for search
}
