% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_table.R
\name{mortalitytableq}
\alias{mortalitytableq}
\title{Getting the mortality rates from a mortality table}
\usage{
mortalitytableq(tab, age, dur)
}
\arguments{
\item{tab}{A table generated by `mortalitytable` function.}

\item{age}{staring age}

\item{dur}{duration}
}
\value{
A list contains various information in the table you have asked
}
\description{
Access to mortality rate in a mortality table
}
